+++ plugins/approvej-intellij-plugin/src/main/kotlin/org/approvej/intellij/ApproveJDiffExtension.kt
@@ -4,6 +4,7 @@ import com.intellij.diff.DiffContext
 import com.intellij.diff.DiffExtension
 import com.intellij.diff.FrameDiffTool
 import com.intellij.diff.requests.DiffRequest
+import com.intellij.openapi.fileEditor.FileEditorManager
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.psi.PsiMethod
@@ -23,21 +24,26 @@ class ApproveJDiffExtension : DiffExtension() {
     context: DiffContext,
     request: DiffRequest,
   ) {
+    if (viewer is ImageDiffTool.ImageDiffViewer) return
     val receivedFile = request.getUserData(ReceivedFileUtil.RECEIVED_FILE_KEY) ?: return
     val approvedFile = request.getUserData(ReceivedFileUtil.APPROVED_FILE_KEY) ?: return
     val project = context.project ?: return
 
     val testMethod = InventoryUtil.findTestMethod(approvedFile, project)
-    val panel = createNotificationPanel(project, receivedFile, approvedFile, testMethod)
+    val panel =
+      createNotificationPanel(project, receivedFile, approvedFile, testMethod) {
+        FileEditorManager.getInstance(project).openFile(approvedFile, true)
+      }
     addToViewerNotifications(viewer, panel)
   }
 
   companion object {
-    private fun createNotificationPanel(
+    internal fun createNotificationPanel(
       project: Project,
       receivedFile: VirtualFile,
       approvedFile: VirtualFile,
       testMethod: PsiMethod?,
+      afterAction: (() -> Unit)? = null,
     ): JComponent {
       val panel = EditorNotificationPanel(EditorNotificationPanel.Status.Info)
       panel.text = "ApproveJ: Received vs. Approved"
@@ -46,8 +52,12 @@ class ApproveJDiffExtension : DiffExtension() {
       }
       panel.createActionLabel("Approve") {
         ReceivedFileUtil.approve(project, receivedFile, approvedFile)
+        afterAction?.invoke()
       }
-      panel.createActionLabel("Reject") { ReceivedFileUtil.reject(project, receivedFile) }
+      panel.createActionLabel("Reject") {
+        ReceivedFileUtil.reject(project, receivedFile)
+        afterAction?.invoke()
+      }
       return panel
     }